moved Divider to correct directory
